[SiteIsolation] RemoteDOMWindow::setLocation should have same security checks as LocalDOMWindow::setLocation https://bugs.webkit.org/show_bug.cgi?id=296453 rdar://156641216
Reviewed by NOBODY (OOPS!).

Move a number of functions from LocalDOMWindow to DOMWindow and invoke them in a new setLocationSecurityChecks() function which is called in both LocalDOMWindow and RemoteDOMWindow.

* Source/WebCore/page/DOMWindow.cpp:
(WebCore::DOMWindow::isCurrentlyDisplayedInFrame const):
(WebCore::DOMWindow::isInsecureScriptAccess):
(WebCore::DOMWindow::crossDomainAccessErrorMessage):
(WebCore::DOMWindow::printErrorMessage const):
(WebCore::DOMWindow::setLocationSecurityChecks):
* Source/WebCore/page/DOMWindow.h:
* Source/WebCore/page/LocalDOMWindow.cpp:
(WebCore::LocalDOMWindow::setLocation):
(WebCore::LocalDOMWindow::isCurrentlyDisplayedInFrame const): Deleted.
(WebCore::LocalDOMWindow::printErrorMessage const): Deleted.
(WebCore::LocalDOMWindow::crossDomainAccessErrorMessage): Deleted.
(WebCore::LocalDOMWindow::isInsecureScriptAccess): Deleted.
* Source/WebCore/page/LocalDOMWindow.h:
* Source/WebCore/page/RemoteDOMWindow.cpp:
(WebCore::RemoteDOMWindow::setLocation):
